// readflex-frontend-master/src/app/shared/models/story-categories.enum.ts
export enum StoryCategories {
  ContesLegendes = 'Contes-et-Légendes',
  AventuresEpiques = 'Aventures Épiques',
  Fantastique = 'Fantastique',
  DrameEmotionnel = 'Drame Émotionnel',
  ScienceFiction = 'Science-fiction et Dystopie',
  HistoireAmour = 'Histoires d’Amour',
  Thriller = 'Suspense et Thriller',
  Enfants = 'Histoires pour Enfants',
  Historique = 'Récits Historiques',
  Inspiration = 'Inspiration et Développement Personnel',
  Horreur = 'Horreur et Surnaturel',
  Comedie = 'Comédie et Humour',
}